refactor(storage): split the SQLite values grab-bag by dependency
`values.ts` held three unrelated concerns behind a name that described
none of them: JSON codecs, `spaces` row statements, and collision-key
allocation.

Split it where a dependency boundary already ran. `rows.ts` owns
everything that touches a stored column or `DatabaseSync`. `identity.ts`
owns the pure title and label allocation rules, and imports no SQLite at
all — which is what makes the cut a boundary rather than a preference.

Several consumers now depend on less: `space-logs`, `space-tasks` and
`structured-store` need only `rows.ts`, and `space-nodes` takes a single
symbol from `identity.ts` in place of a mixed five-symbol block.
